backend/pyproject.toml
[tool.poetry]
name = "backend-kafka-project"
version = "0.1.0"
description = ""
authors = ["Bogdan Babiychuk <babiychuk.bogdan@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
motor = ">=3.7.1,<4.0.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic-settings = ">=2.10.1,<3.0.0"
punq = "^0.7.0"
aiokafka = "^0.12.0"
orjson = "^3.11.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
